requirements.txt
requests[socks]==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
PyYAML==6.0.2
networkx==3.4.2
pandas==2.2.3
pyvis==0.3.2
tqdm==4.67.1